requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
google-cloud-firestore==2.13.1
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
redis==5.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
